// app/backend/src/interfaces/IBoard.ts
export interface IBoard {
  name: string;
  totalPoints: number;
  totalGames: number;
  totalVictories: number;
  totalDraws: number;
  totalLosses: number;
  goalsFavor: number;
  goalsOwn: number;
  goalsBalance: number;
  efficiency: number;
}

export interface ITotalPoints {
  homeTeam: number;
  result: number;
}